ai-sql-tool/requirements.txt
streamlit>=1.35.0
groq>=0.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
plotly>=5.20.0
pandas>=2.2.0
python-dotenv>=1.0.0
bcrypt>=4.1.2
